Update extensionConfig.ts
As it stands now, the build_log.json file will contain gaps like this:
```
        {
          "test": {},
          "use": "file-loader"
        },
```
because of the way JSON.stringify works on regex types.

This would be a development change only.  To be super safe it probably makes sense to unset (or restore) the RegExp prototype to a clean state?